fix(theme): added a workaround so that defaultProps are also picked up from MUI v5 components (like Button disabledRipple or Grid spacing)
